+++ 2023/09/9.cpp
@@ -1,7 +1,7 @@
 #include "../../AOC.h"
 
-vector<long> getDerivative(const vector<long>& v){
-    vector<long> dv(v.size()-1);
+vector<int> getDerivative(const vector<int>& v){
+    vector<int> dv(v.size()-1);
     for(int i=0; i<dv.size(); ++i) {
         dv[i] = v[i+1] - v[i];
     }
@@ -9,7 +9,7 @@ vector<long> getDerivative(const vector<long>& v){
     return dv;
 }
 
-long next(vector<long> seq) {
+int next(vector<int> seq) {
     if(accumulate(seq.begin(), seq.end(), 0L) == 0L)
         return 0L;
 
@@ -21,13 +21,13 @@ signed main(signed argc, char* argv[]) {
     AoCTimer timer(argc, argv);
     
     string str;
-    long ans = 0;
+    int ans = 0;
 
     while(getline(cin, str)) {
         istringstream iss(str);
-        vector<long> v;
+        vector<int> v;
 
-        long a;
+        int a;
         while(iss >> a)
             v.push_back(a);
